DSA: Correctly determine VID
The 4 MSBs of the VID is stored in the lower nibble of the tag's third
byte.

Previously the priority bits where folded into the VID space, e.g. a
packet with VID=1 and priority 6 was printed as having a VID of
1537 (0x601).

Update the DSA PCAPs with packets containing non-zero VID and FPri
values to make sure we catch any future regressions.

Signed-off-by: Tobias Waldekranz <tobias@waldekranz.com>