ptp: Print majorSdoId field instead of just the first bit
The 4 bits in this location are labelled transportSpecific in
IEEE1588-2008. The first bit was dedicated for v1 compatible hardware,
but in Annex F: "Transport over IEEE 802.3 /Ethernet" it says that a
value of transportSpecific 0x1 is reserved for use by standards
developed by the 802.1 AVB Task Group, which later resulted in the
802.1AS standard which uses that value. So it no longer means only v1
compat, and we should just print the whole value.

802.1AS-2020 also renames this field to majorSdoId, and that is the name
used in Wireshark.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>