HSR/PRP: Add printing for HSR header
HSR is a redundancy protocol that duplicates traffic and sends it two
ways in a ring with an HSR header that includes a sequence number. Other
devices in the ring will forward the first occurrence of a packet, and
discard the duplicate (based on SMAC + sequence number). This enables
zero packet loss when a link goes down.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>