HSR/PRP: Add printing for PRP trailer
PRP is a redundancy protocol, like HSR, that duplicates packets and
includes a sequence number. However, PRP uses a trailer on the packet,
making the parsing a bit tricky. The trailer includes the packet size
that can be used to help ascertain it is a valid PRP trailer, however it
is still possible that this can appear naturally in a packet somewhere
in a network that doesn't use PRP.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>